Dedupe MAD st_common.c
